+++ src/test/java/bowling/domain/player/NameTest.java
@@ -1,4 +1,4 @@
-package bowling.domain.name;
+package bowling.domain.player;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
